+++ src/lib.rs
@@ -1,1 +1,11 @@
+use iced::Point;
+
 pub mod plugin;
+pub mod views;
+
+#[derive(Debug, Clone, Copy)]
+pub enum Message {
+    MouseClick(Point),
+    MouseDrag(Point),
+    MouseRelease,
+}
